requirements.txt
PyMuPDF==1.25.3
google-generativeai==0.8.4
openpyxl==3.1.5
paddleocr==2.10.0
paddlepaddle==2.6.2
pillow==11.1.0
protobuf==3.20.3
python-dotenv==1.0.1
streamlit==1.43.0
